Bug fixes and implementations
Fixed a bug where tap detection for beers was non-functional due to code that was previously thought to be obsolete.

Finished switch to new font

Started implementation of a drunkenness effect. An alternative may need to be found as Unity's post processing package causes major performance loss on mobile devices